+++ sdk/python/pulumi_digitalocean/floating_ip.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 
 __all__ = ['FloatingIp']
@@ -15,7 +15,7 @@ class FloatingIp(pulumi.CustomResource):
     def __init__(__self__,
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
-                 droplet_id: Optional[pulumi.Input[float]] = None,
+                 droplet_id: Optional[pulumi.Input[int]] = None,
                  ip_address: Optional[pulumi.Input[str]] = None,
                  region: Optional[pulumi.Input[str]] = None,
                  __props__=None,
@@ -45,7 +45,7 @@ class FloatingIp(pulumi.CustomResource):
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
+        :param pulumi.Input[int] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
         :param pulumi.Input[str] ip_address: The IP Address of the resource
         :param pulumi.Input[str] region: The region that the Floating IP is reserved to.
         """
@@ -82,7 +82,7 @@ class FloatingIp(pulumi.CustomResource):
     def get(resource_name: str,
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
-            droplet_id: Optional[pulumi.Input[float]] = None,
+            droplet_id: Optional[pulumi.Input[int]] = None,
             floating_ip_urn: Optional[pulumi.Input[str]] = None,
             ip_address: Optional[pulumi.Input[str]] = None,
             region: Optional[pulumi.Input[str]] = None) -> 'FloatingIp':
@@ -93,7 +93,7 @@ class FloatingIp(pulumi.CustomResource):
         :param str resource_name: The unique name of the resulting resource.
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
+        :param pulumi.Input[int] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
         :param pulumi.Input[str] floating_ip_urn: The uniform resource name of the floating ip
         :param pulumi.Input[str] ip_address: The IP Address of the resource
         :param pulumi.Input[str] region: The region that the Floating IP is reserved to.
@@ -110,7 +110,7 @@ class FloatingIp(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="dropletId")
-    def droplet_id(self) -> pulumi.Output[Optional[float]]:
+    def droplet_id(self) -> pulumi.Output[Optional[int]]:
         """
         The ID of Droplet that the Floating IP will be assigned to.
         """
