+++ sdk/python/pulumi_digitalocean/volume_attachment.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 
 __all__ = ['VolumeAttachment']
@@ -15,7 +15,7 @@ class VolumeAttachment(pulumi.CustomResource):
     def __init__(__self__,
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
-                 droplet_id: Optional[pulumi.Input[float]] = None,
+                 droplet_id: Optional[pulumi.Input[int]] = None,
                  volume_id: Optional[pulumi.Input[str]] = None,
                  __props__=None,
                  __name__=None,
@@ -47,7 +47,7 @@ class VolumeAttachment(pulumi.CustomResource):
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] droplet_id: ID of the Droplet to attach the volume to.
+        :param pulumi.Input[int] droplet_id: ID of the Droplet to attach the volume to.
         :param pulumi.Input[str] volume_id: ID of the Volume to be attached to the Droplet.
         """
         if __name__ is not None:
@@ -83,7 +83,7 @@ class VolumeAttachment(pulumi.CustomResource):
     def get(resource_name: str,
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
-            droplet_id: Optional[pulumi.Input[float]] = None,
+            droplet_id: Optional[pulumi.Input[int]] = None,
             volume_id: Optional[pulumi.Input[str]] = None) -> 'VolumeAttachment':
         """
         Get an existing VolumeAttachment resource's state with the given name, id, and optional extra
@@ -92,7 +92,7 @@ class VolumeAttachment(pulumi.CustomResource):
         :param str resource_name: The unique name of the resulting resource.
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] droplet_id: ID of the Droplet to attach the volume to.
+        :param pulumi.Input[int] droplet_id: ID of the Droplet to attach the volume to.
         :param pulumi.Input[str] volume_id: ID of the Volume to be attached to the Droplet.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
@@ -105,7 +105,7 @@ class VolumeAttachment(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="dropletId")
-    def droplet_id(self) -> pulumi.Output[float]:
+    def droplet_id(self) -> pulumi.Output[int]:
         """
         ID of the Droplet to attach the volume to.
         """
